Add files via upload
Q1. Write a program in c to display maximum of two numbers without if statement.
Q2. Write a program in c to display maximum of three numbers using 2 simple if (one condition)
statement.
Q3. Write a program in c to display first 10 Fibonacci numbers.
Q4. Write a program in c to check a number is ARMSTRONG number or not.
Q5. Write a program in c to check a number is twisted prime or not.
Q6. Write a program in c to display PASCAL triangle.
Q7. Write a program in c to display how many digits of a number without loop and if statement.
Q8. Write a program in c to reverse an array.
Q9. Write a program in c to implement sliding window. Ex. (1 2 3 4 5) will be ( 2 1 4 3 5).
Q10. Write a program in c to find the enter number can be generated by adding some number of a
array of numbers. Ex (1 2 3 4 5) 8 can be 1+3+4, 1 cannot be possible.
Q11. Write a program in c to merge two arrays of characters.
Q12. Write a program in c string in present in another string or not.
Q13. Write a program in c to display the frequency of vowels of a string.
Q14. Write a program in c to display unique characters from a string. (Remove duplicate).
Q15. Write a program in c to swap two number using call by address.
Q16. Write a program in c to return maximum and minimum out of three numbers from one function.
Q17. Write a program in c to return maximum when the function called first time and minimum when
the function called second time.
Q18. Write a program in c to create a date structure within a student structure.
Q19. Write a program in c to sort the student structure (name, cgpa) according to cgpa.
Q20. Write a program in c to join the elements of two structures.
Q21. Write a program in c to display name of the student having cgpa &gt; 9.8.
Q22. Write a program in c to display details of student structure by passing to a function.
Q23. Write a program in c to display cgpa of a particular student (by name) by sending structure and
name to a function.

Q24. Write a program in c to display name and cgpa where the name contains “KUMAR” by passing
structure to a function.
Q25. Write a program in c to display name and cgpa where the name contains only one vowel “E”.
Q26. Write a menu driven program to add a data, remove data, and modify data from a structure
(name, cgpa).
Q27. Write a program in c to add money to the account (account no, name, balance), or withdraw
money from the account through a menu driven program.
Q28. Write a program in c to display cgpa from the structure (name, cgpa), and the name can be
determine from the structure (name, roll), the roll number is searching key.
Q29. Write a program in c to display common information of two same type of structure.
Q30. Write a program in c to display rank particular of student from the structure (roll, name,
m1,m2,m3, grade, rank).